moved generateSchema to extension moved getPath to extension
