+++ src/extensions/index.ts
@@ -1,0 +1,3 @@
+// extensions
+export * from './GetPathExtension';
+export * from './GenerateSchemaExtension';
